chore: refactored students notes page into students strands
- removed `Notes.list_student_strands_notes/2` in favor of `LearningContext.list_student_strands/2` (they are not exactly the same function, but the base logic is, and one is replacing the other)
- changed `/student_notes` route to `/student_strands`
- renamed "Home" and "Notes" student menu items to "Report cards" and "Strands" respectively
