+++ lib/lanttern/notes.ex
@@ -66,74 +66,6 @@ defmodule Lanttern.Notes do
   end
 
   @doc """
-  Returns the list of student strand notes based on their report cards.
-
-  The function lists all strands linked to the student report cards, ordering the results
-  by (report card) cycle descending and strand reports position ascending.
-
-  The list is comprised of tuples of note (or `nil`) and strand.
-
-  ### Options:
-
-  `:cycles_ids` – filter results by given cycles
-
-  ## Examples
-
-      iex> list_student_strands_notes(user, opts)
-      [{%Note{}, %Strand{}}, ...]
-
-  """
-  @spec list_student_strands_notes(user :: User.t(), opts :: Keyword.t()) :: [
-          {Note.t() | nil, Strand.t()}
-        ]
-  def list_student_strands_notes(%{current_profile: profile} = _user, opts \\ []) do
-    strand_student_notes_map =
-      from(
-        n in Note,
-        join: snr in assoc(n, :strand_note_relationship),
-        where: n.author_id == ^profile.id,
-        select: {n, snr}
-      )
-      |> Repo.all()
-      |> Enum.map(fn {n, snr} ->
-        {snr.strand_id, n}
-      end)
-      |> Enum.into(%{})
-
-    from(
-      s in Strand,
-      join: sr in assoc(s, :strand_reports),
-      join: rc in assoc(sr, :report_card),
-      join: c in assoc(rc, :school_cycle),
-      as: :cycles,
-      join: src in assoc(rc, :students_report_cards),
-      order_by: [desc: c.end_at, asc: c.start_at, asc: sr.position],
-      where: src.student_id == ^profile.student_id
-    )
-    |> apply_list_student_strands_notes_opts(opts)
-    |> Repo.all()
-    |> Enum.uniq()
-    |> maybe_preload(preloads: [:subjects, :years])
-    |> Enum.map(fn strand ->
-      {Map.get(strand_student_notes_map, strand.id), strand}
-    end)
-  end
-
-  defp apply_list_student_strands_notes_opts(queryable, []), do: queryable
-
-  defp apply_list_student_strands_notes_opts(queryable, [{:cycles_ids, cycles_ids} | opts])
-       when cycles_ids != [] do
-    from(
-      [_s, cycles: c] in queryable,
-      where: c.id in ^cycles_ids
-    )
-    |> apply_list_student_strands_notes_opts(opts)
-  end
-
-  defp apply_list_student_strands_notes_opts(queryable, [_opt | opts]),
-    do: apply_list_student_strands_notes_opts(queryable, opts)
-
-  @doc """
   Returns the list of students of the given classes with their strand notes.
 
   Students are ordered by class name and name.
