+++ test/lanttern/notes_test.exs
@@ -245,123 +245,6 @@ defmodule Lanttern.NotesTest do
       assert expected_2.moment.id == moment_2.id
     end
 
-    test "list_student_strands_notes/2 returns all user strand notes related to students report cards" do
-      author = student_profile_fixture()
-
-      subject_1 = Lanttern.TaxonomyFixtures.subject_fixture()
-      subject_2 = Lanttern.TaxonomyFixtures.subject_fixture()
-      year = Lanttern.TaxonomyFixtures.year_fixture()
-
-      strand_1 =
-        strand_fixture(%{subjects_ids: [subject_1.id, subject_2.id], years_ids: [year.id]})
-
-      strand_2 = strand_fixture()
-      strand_3 = strand_fixture()
-      strand_4 = strand_fixture()
-
-      cycle_2024 =
-        Lanttern.SchoolsFixtures.cycle_fixture(start_at: ~D[2024-01-01], end_at: ~D[2024-12-31])
-
-      cycle_2023 =
-        Lanttern.SchoolsFixtures.cycle_fixture(start_at: ~D[2023-01-01], end_at: ~D[2023-12-31])
-
-      report_card_2024 =
-        Lanttern.ReportingFixtures.report_card_fixture(%{school_cycle_id: cycle_2024.id})
-
-      report_card_2023 =
-        Lanttern.ReportingFixtures.report_card_fixture(%{school_cycle_id: cycle_2023.id})
-
-      # create strand reports
-
-      _strand_report_1_2024 =
-        Lanttern.ReportingFixtures.strand_report_fixture(%{
-          report_card_id: report_card_2024.id,
-          strand_id: strand_1.id
-        })
-
-      _strand_report_2_2024 =
-        Lanttern.ReportingFixtures.strand_report_fixture(%{
-          report_card_id: report_card_2024.id,
-          strand_id: strand_2.id
-        })
-
-      _strand_report_3_2023 =
-        Lanttern.ReportingFixtures.strand_report_fixture(%{
-          report_card_id: report_card_2023.id,
-          strand_id: strand_3.id
-        })
-
-      _strand_report_4_2023 =
-        Lanttern.ReportingFixtures.strand_report_fixture(%{
-          report_card_id: report_card_2023.id,
-          strand_id: strand_4.id
-        })
-
-      # create students report cards
-      _ =
-        Lanttern.ReportingFixtures.student_report_card_fixture(%{
-          student_id: author.student_id,
-          report_card_id: report_card_2024.id
-        })
-
-      _ =
-        Lanttern.ReportingFixtures.student_report_card_fixture(%{
-          student_id: author.student_id,
-          report_card_id: report_card_2023.id
-        })
-
-      # create student notes in strand 1 and 3
-
-      strand_1_note = strand_note_fixture(%{current_profile: author}, strand_1.id)
-      strand_3_note = strand_note_fixture(%{current_profile: author}, strand_3.id)
-
-      # extra fixtures for filter testing
-      other_strand = strand_fixture()
-      other_report_card = Lanttern.ReportingFixtures.report_card_fixture()
-
-      _other_strand_1_note =
-        strand_note_fixture(%{current_profile: student_profile_fixture()}, strand_1.id)
-
-      _other_strand_2_note =
-        strand_note_fixture(%{current_profile: student_profile_fixture()}, strand_2.id)
-
-      _other_strand_report =
-        Lanttern.ReportingFixtures.strand_report_fixture(%{
-          report_card_id: other_report_card.id,
-          strand_id: other_strand.id
-        })
-
-      _ =
-        Lanttern.ReportingFixtures.student_report_card_fixture(%{
-          student_id: author.student_id,
-          report_card_id: other_report_card.id
-        })
-
-      assert [
-               {^strand_1_note, expected_strand_1},
-               {nil, expected_strand_2},
-               {^strand_3_note, expected_strand_3},
-               {nil, expected_strand_4}
-             ] =
-               Notes.list_student_strands_notes(
-                 %{current_profile: author},
-                 cycles_ids: [cycle_2023.id, cycle_2024.id]
-               )
-
-      # assert expected_strand_1_note.id == strand_1_note.id
-      assert expected_strand_1.id == strand_1.id
-      assert subject_1 in expected_strand_1.subjects
-      assert subject_2 in expected_strand_1.subjects
-      assert [year] == expected_strand_1.years
-
-      assert expected_strand_2.id == strand_2.id
-
-      # assert expected_strand_3_note.id == strand_3_note.id
-      assert expected_strand_3.id == strand_3.id
-
-      assert expected_strand_4.id == strand_4.id
-    end
-
     test "list_classes_strand_notes/2 returns the list of classes students and their strand notes" do
       class_a = Lanttern.SchoolsFixtures.class_fixture(%{name: "A"})
       class_b = Lanttern.SchoolsFixtures.class_fixture(%{name: "B"})
